+++ CHANGES.md
@@ -4,6 +4,8 @@
 - Deprecated old decorators, added `@Properties`
 - `$inflate` and `$deflate` now properly behave with nested components
 - Added `ComponentStore`, which is added to when `@Properties` is used
+- `Component#$locate` is now recursive
+- Classes are now properly added in `Component#$ensureElement`
 - Removed `Object.assign` requirements, now using `object-assign`
 
 ## 0.1.4